fix(huobijp): rename into bittrade feat(bittrade): rename from huobijp
