+++ ts/src/abstract/bittrade.ts
